requirements.txt
# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
alembic==1.12.1

# NLP Libraries
nltk==3.8.1
spacy==3.7.2
transformers==4.35.2
torch==2.1.1
sentence-transformers==2.2.2
langdetect==1.0.9

# Computer Vision & OCR
opencv-python==4.8.1.78
Pillow==10.1.0
pytesseract==0.3.10
imagehash==4.3.1

# Web Scraping
beautifulsoup4==4.12.2
scrapy==2.11.0
requests==2.31.0
lxml==4.9.3
selenium==4.15.2
webdriver-manager==4.0.1

# WhatsApp Automation
pywhatkit==5.4
pyautogui==0.9.54

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3
schedule==1.2.0

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Logging & Monitoring
python-json-logger==2.0.7
colorlog==6.8.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1

# Security
cryptography==41.0.7

# Rate Limiting
Flask-Limiter==3.5.0
slowapi==0.1.9
